Refactor to moved duplicated code in AsyncXXXLike traits to AsyncTests.scala.
