+++ scalatest/src/main/scala/org/scalatest/AsyncTests.scala
@@ -15,12 +15,63 @@
  */
 package org.scalatest
 
-import scala.concurrent.Future
+import scala.concurrent.{ExecutionContext, Future}
 
-trait AsyncTests extends SuiteMixin with AsyncFixtures { this: Suite with TestRegistration =>
+trait AsyncTests extends SuiteMixin with AsyncFixtures { thisSuite: Suite with TestRegistration with OneInstancePerTest =>
   type Registration = Future[Unit]
 
   import scala.language.implicitConversions
 
   implicit def convertToFuture(o: Any): Future[Unit] = Future.successful(o)
+
+  implicit def executionContext: ExecutionContext
+
+  override private[scalatest] def transformToOutcome(testFun: => Future[Unit]): () => AsyncOutcome =
+    () => {
+      val futureUnit = testFun
+      FutureOutcome(
+        futureUnit.map(u => Succeeded).recover {
+          case ex: exceptions.TestCanceledException => Canceled(ex)
+          case _: exceptions.TestPendingException => Pending
+          case tfe: exceptions.TestFailedException => Failed(tfe)
+          case ex: Throwable if !Suite.anExceptionThatShouldCauseAnAbort(ex) => Failed(ex)
+        }
+      )
+    }
+
+  protected[scalatest] def getEngine: Engine
+
+  private final val engine = getEngine
+
+  import engine._
+
+  protected override def runTest(testName: String, args: Args): Status = {
+
+    if (args.runTestInNewInstance) {
+      // In initial instance, so create a new test-specific instance for this test and invoke run on it.
+      val oneInstance = newInstance
+      oneInstance.run(Some(testName), args)
+    }
+    else {
+      // Therefore, in test-specific instance, so run the test.
+      def invokeWithAsyncFixture(theTest: TestLeaf): AsyncOutcome = {
+        val theConfigMap = args.configMap
+        val testData = testDataFor(testName, theConfigMap)
+        FutureOutcome(
+          withAsyncFixture(
+            new NoArgAsyncTest {
+              val name = testData.name
+              def apply(): Future[Outcome] = { theTest.testFun().toFutureOutcome }
+              val configMap = testData.configMap
+              val scopes = testData.scopes
+              val text = testData.text
+              val tags = testData.tags
+            }
+          )
+        )
+      }
+
+      runTestImpl(thisSuite, testName, args, true, invokeWithAsyncFixture)
+    }
+  }
 }
