+++ scalatest/src/main/scala/org/scalatest/AsyncWordSpecLike.scala
@@ -18,54 +18,4 @@ package org.scalatest
 import scala.concurrent.{ExecutionContext, Future}
 
 //SCALATESTJS-ONLY @scala.scalajs.js.annotation.JSExportDescendentClasses(ignoreInvalidDescendants = true)
-trait AsyncWordSpecLike extends WordSpecRegistration with AsyncTests with OneInstancePerTest { thisSuite =>
-
-  implicit def executionContext: ExecutionContext
-
-  override private[scalatest] def transformToOutcome(testFun: => Future[Unit]): () => AsyncOutcome =
-    () => {
-      val futureUnit = testFun
-      FutureOutcome(
-        futureUnit.map(u => Succeeded).recover {
-          case ex: exceptions.TestCanceledException => Canceled(ex)
-          case _: exceptions.TestPendingException => Pending
-          case tfe: exceptions.TestFailedException => Failed(tfe)
-          case ex: Throwable if !Suite.anExceptionThatShouldCauseAnAbort(ex) => Failed(ex)
-        }
-      )
-    }
-
-  private final val engine: Engine = getEngine
-
-  import engine._
-
-  protected override def runTest(testName: String, args: Args): Status = {
-
-    if (args.runTestInNewInstance) {
-      // In initial instance, so create a new test-specific instance for this test and invoke run on it.
-      val oneInstance = newInstance
-      oneInstance.run(Some(testName), args)
-    }
-    else {
-      // Therefore, in test-specific instance, so run the test.
-      def invokeWithAsyncFixture(theTest: TestLeaf): AsyncOutcome = {
-        val theConfigMap = args.configMap
-        val testData = testDataFor(testName, theConfigMap)
-        FutureOutcome(
-          withAsyncFixture(
-            new NoArgAsyncTest {
-              val name = testData.name
-              def apply(): Future[Outcome] = { theTest.testFun().toFutureOutcome }
-              val configMap = testData.configMap
-              val scopes = testData.scopes
-              val text = testData.text
-              val tags = testData.tags
-            }
-          )
-        )
-      }
-
-      runTestImpl(thisSuite, testName, args, true, invokeWithAsyncFixture)
-    }
-  }
-}
+trait AsyncWordSpecLike extends WordSpecRegistration with AsyncTests with OneInstancePerTest
